Auth plus swagger

+++ AppNiZiAPI/Functions/DietaryManagement/GET/DietaryManagement.cs
@@ -38,7 +38,7 @@ namespace AppNiZiAPI.Functions.DietaryManagement.GET
 
             AuthResultModel authResult = await DIContainer.Instance.GetService<IAuthorization>().CheckAuthorization(req, patientId);
             if (!authResult.Result)
-                return new StatusCodeResult(authResult.StatusCode);
+                return new StatusCodeResult((int)authResult.StatusCode);
 
             List<DietaryManagementModel> dietaryManagementModels;
             List<DietaryRestriction> dietaryRestrictions;
