+++ AppNiZiAPI/Functions/Patients/DELETE/DeletePatient.cs
@@ -30,7 +30,7 @@ namespace AppNiZiAPI.Functions.Patients.DELETE
             #region AuthCheck
             AuthResultModel authResult = await DIContainer.Instance.GetService<IAuthorization>().CheckAuthorization(req, patientId);
             if (!authResult.Result)
-                return new StatusCodeResult(authResult.StatusCode);
+                return new StatusCodeResult((int)authResult.StatusCode);
             #endregion
 
             if (patientId == 0)
